+++ internal/bot/bot.go
@@ -14,6 +14,7 @@ import (
 	"github.com/codemicro/lgballtDiscordBot/internal/bot/components/muteme"
 	"github.com/codemicro/lgballtDiscordBot/internal/bot/components/pressf"
 	"github.com/codemicro/lgballtDiscordBot/internal/bot/components/roles"
+	"github.com/codemicro/lgballtDiscordBot/internal/bot/components/toneTags"
 	"github.com/codemicro/lgballtDiscordBot/internal/bot/components/verification"
 	"github.com/codemicro/lgballtDiscordBot/internal/buildInfo"
 	"github.com/codemicro/lgballtDiscordBot/internal/config"
@@ -128,6 +129,11 @@ func registerHandlers(kit *route.Kit, st *state.State) error {
 	}
 
 	err = verification.Init(kit, st)
+	if err != nil {
+		return err
+	}
+
+	err = toneTags.Init(kit, st)
 	if err != nil {
 		return err
 	}