+++ .github/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 (Dates are in YYYY-MM-DD format. This message is mainly for my own sake.)
 
 ## [Unreleased]
+### Added
+* Tone tag lookup commands
+  * `<p>toneTag lookup <shorthand>`, `<p>toneTag list`, `<p>toneTag new <shorthand> <description>`
+  * Also adds `toV47.py` migration script to automatically fill the database with a set of tone tags prior to first use
 ### Changed
 * The previous incident counter streak is now shown when the `<p>indicents reset` command is run