+++ .github/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Tone tag lookup commands
   * `<p>toneTag lookup <shorthand>`, `<p>toneTag list`, `<p>toneTag new <shorthand> <description>`
   * Also adds `toV47.py` migration script to automatically fill the database with a set of tone tags prior to first use
+* `<p>uwu`
 ### Changed
 * The previous incident counter streak is now shown when the `<p>indicents reset` command is run
 * Improved chatchart chart and response formatting